refactor(nucleus-node): extract test module to tests_main.rs
Move 66-line inline test module from main.rs to separate file.
main.rs: 3302 → 3237 lines (65 line reduction, headroom below 3304 ratchet).

All 66 node tests pass identically.
