+++ crates/nucleus-node/src/main.rs
@@ -3232,71 +3232,5 @@ fn pod_info_to_grpc(info: PodInfo) -> proto::PodInfo {
 // Firecracker VM configuration lives in firecracker_config.rs
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn label_selector_empty_matches_all() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(matches_label_selector(&labels, ""));
-    }
-
-    #[test]
-    fn label_selector_empty_labels_no_match() {
-        let labels = BTreeMap::new();
-        assert!(!matches_label_selector(&labels, "team=backend"));
-    }
-
-    #[test]
-    fn label_selector_single_match() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(matches_label_selector(&labels, "team=backend"));
-    }
-
-    #[test]
-    fn label_selector_single_no_match() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(!matches_label_selector(&labels, "team=frontend"));
-    }
-
-    #[test]
-    fn label_selector_multiple_and_semantics() {
-        let labels = BTreeMap::from([
-            ("team".into(), "backend".into()),
-            ("env".into(), "prod".into()),
-        ]);
-        assert!(matches_label_selector(&labels, "team=backend,env=prod"));
-        assert!(!matches_label_selector(&labels, "team=backend,env=staging"));
-    }
-
-    #[test]
-    fn label_selector_whitespace_trimmed() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(matches_label_selector(&labels, " team = backend "));
-    }
-
-    #[test]
-    fn label_selector_missing_value_no_match() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(!matches_label_selector(&labels, "team"));
-    }
-
-    #[test]
-    fn label_selector_missing_key_in_labels() {
-        let labels = BTreeMap::from([("team".into(), "backend".into())]);
-        assert!(!matches_label_selector(&labels, "env=prod"));
-    }
-
-    #[test]
-    fn label_selector_value_with_equals_sign() {
-        // key=val=ue should parse as key="val=ue" thanks to splitn(2, '=')
-        let labels = BTreeMap::from([("expr".into(), "a=b".into())]);
-        assert!(matches_label_selector(&labels, "expr=a=b"));
-    }
-
-    #[test]
-    fn label_selector_empty_value() {
-        let labels = BTreeMap::from([("tag".into(), "".into())]);
-        assert!(matches_label_selector(&labels, "tag="));
-    }
-}
+#[path = "tests_main.rs"]
+mod tests;
